SILKIT-1220: make data and rpc info available for dashboard (#552)
for data publisher and subscriber: topic, media type and labels
for data subscriber internal: data publisher uuid
for rpc client and server: function name, media type and labels
for rpc server internal: rpc client uuid

+++ SilKit/source/dashboard/Client/Test_DashboardSystemServiceClient.cpp
@@ -185,7 +185,7 @@ TEST_F(TestDashboardSystemServiceClient, SetSimulationEnd_Success)
 
     // Act
     const auto service = CreateService();
-    const oatpp::UInt32 expectedSimulationId = 123;
+    const oatpp::UInt64 expectedSimulationId = 123;
     auto request = SimulationEndDto::createShared();
     service->SetSimulationEnd(expectedSimulationId, request);
     WaitForExecutor();
@@ -212,7 +212,7 @@ TEST_F(TestDashboardSystemServiceClient, SetSimulationEnd_Failure)
 
     // Act
     const auto service = CreateService();
-    const oatpp::UInt32 expectedSimulationId = 123;
+    const oatpp::UInt64 expectedSimulationId = 123;
     auto request = SimulationEndDto::createShared();
     service->SetSimulationEnd(expectedSimulationId, request);
     WaitForExecutor();
